Feature/nose wheel control (#169)
* Use std::abs() instead of qAbs()

* Add static_cast<int>

* Update geographiclib to v2.4 (44a5c2d)

* Update cpptrace to v0.6.3 (90de25f)

* Reset replay mode to "normal" when restoring a flight

- Properly initialse replay mode again when switching to the formation
module
- FIX: Do to initialise user aircraft position upon replay mode changes
when no flight is currently loaded ("no teleportation to the equator")

* Add nose wheel steering support

* Correct nose wheel steering direction

+++ src/Plugins/Module/Logbook/src/LogbookWidget.cpp
@@ -180,19 +180,19 @@ void LogbookWidget::initUi() noexcept
         tr("Arrival"),
         tr("Total Time of Flight")
     };
-    LogbookWidgetPrivate::flightIdColumn = headers.indexOf(tr("Flight"));
-    LogbookWidgetPrivate::titleColumn = headers.indexOf(tr("Title"));
-    LogbookWidgetPrivate::flightNumberColumn = headers.indexOf(tr("Flight Number"));
-    LogbookWidgetPrivate::userAircraftColumn = headers.indexOf(tr("User Aircraft"));    
-    LogbookWidgetPrivate::aircraftCountColumn = headers.indexOf(tr("Number of Aircraft"));
-    LogbookWidgetPrivate::recordingDateColumn = headers.indexOf(tr("Recording Date"));
-    LogbookWidgetPrivate::startTimeColumn = headers.indexOf(tr("Departure Time"));
-    LogbookWidgetPrivate::startLocationColumn = headers.indexOf(tr("Departure"));
-    LogbookWidgetPrivate::endTimeColumn = headers.indexOf(tr("Arrival Time"));
-    LogbookWidgetPrivate::endLocationColumn = headers.indexOf(tr("Arrival"));
-    LogbookWidgetPrivate::durationColumn = headers.indexOf(tr("Total Time of Flight"));
+    LogbookWidgetPrivate::flightIdColumn = static_cast<int>(headers.indexOf(tr("Flight")));
+    LogbookWidgetPrivate::titleColumn = static_cast<int>(headers.indexOf(tr("Title")));
+    LogbookWidgetPrivate::flightNumberColumn = static_cast<int>(headers.indexOf(tr("Flight Number")));
+    LogbookWidgetPrivate::userAircraftColumn = static_cast<int>(headers.indexOf(tr("User Aircraft")));
+    LogbookWidgetPrivate::aircraftCountColumn = static_cast<int>(headers.indexOf(tr("Number of Aircraft")));
+    LogbookWidgetPrivate::recordingDateColumn = static_cast<int>(headers.indexOf(tr("Recording Date")));
+    LogbookWidgetPrivate::startTimeColumn = static_cast<int>(headers.indexOf(tr("Departure Time")));
+    LogbookWidgetPrivate::startLocationColumn = static_cast<int>(headers.indexOf(tr("Departure")));
+    LogbookWidgetPrivate::endTimeColumn = static_cast<int>(headers.indexOf(tr("Arrival Time")));
+    LogbookWidgetPrivate::endLocationColumn = static_cast<int>(headers.indexOf(tr("Arrival")));
+    LogbookWidgetPrivate::durationColumn = static_cast<int>(headers.indexOf(tr("Total Time of Flight")));
 
-    ui->logTableWidget->setColumnCount(headers.count());
+    ui->logTableWidget->setColumnCount(static_cast<int>(headers.count()));
     ui->logTableWidget->setHorizontalHeaderLabels(headers);
     ui->logTableWidget->setSelectionBehavior(QAbstractItemView::SelectRows);
     ui->logTableWidget->setSelectionMode(QAbstractItemView::SingleSelection);
